requirements.txt
streamlit
qdrant-client
ctransformers
sentence-transformers
transformers
pypdf
python-docx
numpy
scipy
tqdm
loguru
PyYAML
psutil
requests 